feat(report_generator): rename atta to report_generator
