+++ MANIFEST.in
@@ -1,1 +1,2 @@
-include atta/data/*
+include report_generator/data/*
+include report_generator/exporter/data/*
