+++ report_generator/exporter/assets/favicon-16x16.png
