+++ report_generator/exporter/assets/favicon-32x32.png
