+++ report_generator/exporter/assets/favicon.ico
